More refactoring and removal of duplication (especially in tests).
