+++ pcgrandom/test/test_pcg_xsl_rr_v0.py
@@ -3,13 +3,12 @@ Tests for the PCG_XSL_RR_V0 generator.
 """
 from __future__ import division
 
-import collections
 import contextlib
-import itertools
 import math
 import unittest
 
 from pcgrandom.pcg_xsl_rr_v0 import PCG_XSL_RR_V0
+from pcgrandom.test.test_common import TestCommon
 
 
 # Sequences used for detecting reproducibility regressions.
@@ -74,7 +73,11 @@ def count_samples_generated(gen):
         del gen._next_word
 
 
-class Test_PCG_XSL_RR_V0(unittest.TestCase):
+class Test_PCG_XSL_RR_V0(TestCommon, unittest.TestCase):
+
+    def setUp(self):
+        self.gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
+
     def test_creation_without_seed(self):
         gen1 = PCG_XSL_RR_V0()
         gen2 = PCG_XSL_RR_V0()
@@ -224,54 +227,6 @@ class Test_PCG_XSL_RR_V0(unittest.TestCase):
         samples2 = [gen.gauss(0.0, 1.0) for _ in range(5)]
         self.assertEqual(samples1, samples2)
 
-    def test_getrandbits(self):
-        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
-
-        k = 5
-        samples = [gen.getrandbits(k) for _ in range(10000)]
-        self.check_uniform(range(2**k), samples)
-
-    def test_getrandbits_large(self):
-        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
-
-        k = 379
-        samples = [gen.getrandbits(k) for _ in range(1000)]
-
-        self.assertGreaterEqual(min(samples), 0)
-        self.assertLess(max(samples), 2**379)
-        # It's possible, but vanishingly unlikely, that all values
-        # are less than 2**378.
-        self.assertGreaterEqual(max(samples), 2**378)
-
-    def test_getrandbits_zero_bits(self):
-        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
-        result = gen.getrandbits(0)
-        self.assertEqual(result, 0)
-
-    def test_getrandbits_negative_bits(self):
-        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
-        with self.assertRaises(ValueError):
-            gen.getrandbits(-1)
-
-    def test_randrange_uniform(self):
-        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
-
-        # Indirect test of our _randbelow override.
-        n = 13
-        samples = [gen.randrange(n) for _ in range(10000)]
-        self.check_uniform(range(n), samples)
-
-    def test_randrange_one(self):
-        # Corner case.
-        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
-        state_before = gen.getstate()
-
-        samples = [gen.randrange(1) for _ in range(10)]
-        self.assertEqual(samples, [0]*10)
-
-        state_after = gen.getstate()
-        self.assertEqual(state_before, state_after)
-
     def test_randrange_of_wordsize(self):
         # Using randrange(2**64) should consume exactly one
         # sample each time.
@@ -312,213 +267,6 @@ class Test_PCG_XSL_RR_V0(unittest.TestCase):
             self.assertEqual(sample, samples[next_pos])
             current_pos = next_pos + 1
 
-    def test_randrange_float_arguments(self):
-        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
-
-        with self.assertRaises(TypeError):
-            gen.randrange(5.0)
-        with self.assertRaises(TypeError):
-            gen.randrange(2.0, 7)
-        with self.assertRaises(TypeError):
-            gen.randrange(2, 7.0)
-        with self.assertRaises(TypeError):
-            gen.randrange(2, 7, 1.0)
-        with self.assertRaises(TypeError):
-            gen.randrange(2, 7, 2.0)
-
-    def test_randrange_start_only(self):
-        # See http://bugs.python.org/issue9379 for related discussion.
-        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
-        samples = [gen.randrange(23) for _ in range(1000)]
-        # Chance of getting only 22 of the possible 23 outcomes is < 1.2e-18.
-        self.assertEqual(set(samples), set(range(23)))
-
-        with self.assertRaises(ValueError):
-            gen.randrange(0)
-        with self.assertRaises(ValueError):
-            gen.randrange(-5)
-
-    def test_randrange_start_and_stop(self):
-        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
-        samples = [gen.randrange(-10, 13) for _ in range(1000)]
-        self.assertEqual(set(samples), set(range(-10, 13)))
-
-        with self.assertRaises(ValueError):
-            gen.randrange(20, 20)
-        with self.assertRaises(ValueError):
-            gen.randrange(21, 20)
-
-    def test_randrange_start_stop_and_step(self):
-        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
-
-        test_ranges = [
-            (16, 37, 5),
-            (5, 35, 5),
-            (4, 35, 5),
-            (35, 5, -5),
-            (36, 5, -5),
-        ]
-
-        for test_range in test_ranges:
-            samples = [gen.randrange(*test_range) for _ in range(1000)]
-            self.assertEqual(set(samples), set(range(*test_range)))
-
-        with self.assertRaises(ValueError):
-            gen.randrange(0, 20, 0)
-        with self.assertRaises(ValueError):
-            gen.randrange(0, -20, 0)
-        with self.assertRaises(ValueError):
-            gen.randrange(0, 5, -1)
-        with self.assertRaises(ValueError):
-            gen.randrange(0, -5, 1)
-        with self.assertRaises(ValueError):
-            gen.randrange(22, 22, -1)
-        with self.assertRaises(ValueError):
-            gen.randrange(47, 47, 1)
-
-    def test_choice(self):
-        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
-
-        with self.assertRaises(IndexError):
-            gen.choice([])
-
-        nsamples = 1000
-        seq = "ABCD"
-        choices = [
-            gen.choice(seq) for _ in range(nsamples)
-        ]
-        self.check_uniform(seq, choices)
-
-    def test_sample(self):
-        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
-
-        samples = [tuple(gen.sample(range(4), 3)) for _ in range(10000)]
-        population = list(itertools.permutations(range(4), 3))
-        self.check_uniform(population, samples)
-
-    def test_sample_set(self):
-        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
-
-        s = set('ABCDEFG')
-        sample = gen.sample(s, 3)
-        self.assertLess(set(sample), s)
-
-    def test_sample_dict(self):
-        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
-
-        d = dict(a=1, b=2, c=3)
-        with self.assertRaises(TypeError):
-            gen.sample(d, 2)
-
-    def test_sample_corner_cases(self):
-        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
-        self.assertEqual(gen.sample([], 0), [])
-        with self.assertRaises(ValueError):
-            self.assertEqual(gen.sample([], 1))
-        with self.assertRaises(ValueError):
-            self.assertEqual(gen.sample([], -1))
-
-    def test_shuffle(self):
-        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
-        population = list(range(13))
-
-        result = gen.shuffle(population)
-        self.assertIsNone(result)
-        self.assertNotEqual(population, list(range(13)))
-        self.assertEqual(set(population), set(range(13)))
-
-        population = list(range(4))
-
-        samples = []
-        for _ in range(10000):
-            gen.shuffle(population)
-            samples.append(tuple(population))
-
-        self.check_uniform(
-            list(itertools.permutations(range(4))),
-            samples,
-        )
-
-    def test_shuffle_corner_cases(self):
-        # shuffling a length 0 or 1 sequence shouldn't be a problem
-        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
-        population = []
-        gen.shuffle(population)
-        self.assertEqual(population, [])
-
-        population = [13]
-        gen.shuffle(population)
-        self.assertEqual(population, [13])
-
-    def test_choices(self):
-        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
-
-        population = list(range(5))
-
-        sample = gen.choices(population, k=10000)
-        self.check_uniform(population, sample)
-
-        weights = [2, 3, 0, 1, 4]
-        sample = gen.choices(population, weights=weights, k=10000)
-        self.check_goodness_of_fit(dict(zip(population, weights)), sample)
-
-        cum_weights = [2, 5, 5, 6, 10]
-        sample = gen.choices(population, cum_weights=cum_weights, k=10000)
-        self.check_goodness_of_fit(dict(zip(population, weights)), sample)
-
-    def test_choices_subnormal_weights(self):
-        # Corner case where random.Random triggers an IndexError.
-        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
-
-        population = list(range(5))
-        weights = [1e-323, 0.0, 2e-323, 2e-323, 1e-323]
-        sample = gen.choices(population, weights=weights, k=10000)
-        self.check_goodness_of_fit(dict(zip(population, weights)), sample)
-
-        population = list(range(6))
-        weights = [1e-323, 0.0, 2e-323, 2e-323, 1e-323, 0.0]
-        sample = gen.choices(population, weights=weights, k=10000)
-        self.check_goodness_of_fit(dict(zip(population, weights)), sample)
-
-    def test_choices_error_conditions(self):
-        gen = PCG_XSL_RR_V0(seed=15206, sequence=1729)
-
-        # Can't specify both weights and cum_weights.
-        with self.assertRaises(TypeError):
-            gen.choices(range(3), weights=[1, 2, 3], cum_weights=[1, 3, 6])
-
-        # Two errors here: empty population, and both weights and cum_weights
-        # specified. The TypeError for the bad signature should win.
-        with self.assertRaises(TypeError):
-            gen.choices([], weights=[1, 2, 3], cum_weights=[1, 3, 6])
-
-        # Empty population: IndexError to match random.choice. We should get an
-        # exception even in the corner case that zero samples are requested.
-        with self.assertRaises(IndexError):
-            gen.choices([], weights=[])
-        with self.assertRaises(IndexError):
-            gen.choices([], cum_weights=[])
-        with self.assertRaises(IndexError):
-            gen.choices([])
-        with self.assertRaises(IndexError):
-            gen.choices([], weights=[], k=0)
-        with self.assertRaises(IndexError):
-            gen.choices([], cum_weights=[], k=0)
-        with self.assertRaises(IndexError):
-            gen.choices([], k=0)
-
-        # Number of weights doesn't match population.
-        with self.assertRaises(ValueError):
-            gen.choices(range(3), weights=[1, 2])
-        with self.assertRaises(ValueError):
-            gen.choices(range(3), cum_weights=[1, 2])
-
-        # Total weight is zero.
-        with self.assertRaises(ValueError):
-            gen.choices(range(3), weights=[0.0, 0.0, 0.0])
-        with self.assertRaises(ValueError):
-            gen.choices(range(3), cum_weights=[0.0, 0.0, 0.0])
-
     def test_count_samples_generated(self):
         # This is really a test for our count_samples_generated helper
         # rather than for the PRNG.
@@ -533,48 +281,3 @@ class Test_PCG_XSL_RR_V0(unittest.TestCase):
 
         [gen.randrange(27) for _ in range(13)]
         self.assertEqual(wordgen.call_count, 31)
-
-    def check_uniform(self, population, sample):
-        """
-        Check uniformity via a chi-squared test with p-value 0.99.
-
-        Requires that there's an entry for len(population) - 1
-        in the chisq_99percentile dictionary.
-
-        Parameters
-        ----------
-        population : sequence
-            The population that the samples are drawn from.
-        sample : sequence
-            The generated sample.
-        """
-        weights = {i: 1 for i in population}
-        self.check_goodness_of_fit(weights, sample)
-
-    def check_goodness_of_fit(self, weights, sample):
-        """
-        Perform a chi-squared goodness of fit test.
-
-        Requires that there's an entry for len(expected) - 1
-        in the chisq_99percentile dictionary.
-
-        sample : the sample to be tested
-        weights : mapping from members of the population to their
-            expected weights. The weights need not be normalised.
-        """
-        counts = collections.Counter(sample)
-        # Expected frequencies for this sample.
-        total = sum(weights.values())
-        expected = {
-            i: w / total * len(sample) for i, w in weights.items() if w}
-
-        # Check that we don't have any extraneous objects in our sample.  This
-        # also acts as a check that elements with zero weight haven't been
-        # generated.
-        self.assertLessEqual(set(counts), set(expected))
-
-        stat = sum(
-            (counts[i] - expected[i])**2 / expected[i]
-            for i in expected
-        )
-        self.assertLess(stat, chisq_99percentile[len(expected)-1])
